+++ .github/workflows/build.yml
@@ -33,8 +33,8 @@ jobs:
             -Dsonar.python.version=3
             -Dsonar.python.coverage.reportPaths=main/build/coverage.xml
             -Dsonar.sources=main
-            -Dsonar.exclusions=tests/*
-            -Dsonar.tests=tests
+            -Dsonar.exclusions=main/tests/*
+            -Dsonar.tests=main/tests
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}  # Needed to get PR information, if any
           SONAR_TOKEN: ${{ secrets.SONARCLOUD_TOKEN }}
